+++ scripts/macros.js
@@ -85,6 +85,7 @@ export {alterSelf} from './macros/2024/spells/alterSelf.js';
 export {animalFriendship} from './macros/2024/spells/animalFriendship.js';
 export {animalMessenger} from './macros/2024/spells/animalMessenger.js';
 export {animateDead} from './macros/2024/spells/animateDead.js';
+export {animateObjects} from './macros/2024/spells/animateObjects.js';
 export {antilifeShell, antilifeShellOtherMove} from './macros/2024/spells/antilifeShell.js';
 export {arcaneVigor} from './macros/2024/spells/arcaneVigor.js';
 export {armorOfAgathys} from './macros/2024/spells/armorOfAgathys.js';
@@ -491,6 +492,7 @@ export {deflectEnergy} from './macros/2024/classFeatures/monk/deflectEnergy.js';
 export {disciplinedSurvivor} from './macros/2024/classFeatures/monk/disciplinedSurvivor.js';
 export {perfectFocus} from './macros/2024/classFeatures/monk/perfectFocus.js';
 export {superiorDefense} from './macros/2024/classFeatures/monk/superiorDefense.js';
+export {monkUnarmedStrike} from './macros/2024/classFeatures/monk/monkUnarmedStrike.js';
 // Warrior of the Elements
 export {elementalAttunement, elementalAttunementElementalStrikes} from './macros/2024/classFeatures/monk/warriorOfTheElements/elementalAttunement.js';
 export {elementalBurst} from './macros/2024/classFeatures/monk/warriorOfTheElements/elementalBurst.js';