+++ scripts/macros/2014/spells/greenFlameBlade.js
@@ -29,6 +29,7 @@ async function use({workflow}) {
             });
         }
     }
+    genericUtils.setProperty(weaponData, 'flags.chris-premades.bladeCantrip', workflow.item.name);
     let attackWorkflow = await workflowUtils.syntheticItemDataRoll(weaponData, workflow.actor, [workflow.targets.first()]);
     if (!attackWorkflow) return;
     if (!attackWorkflow.hitTargets.size) return;