+++ scripts/macros/2024/classFeatures/monk/monkUnarmedStrike.js
@@ -1,0 +1,5 @@
+export let monkUnarmedStrike = {
+    name: 'Unarmed Strike (Monk)',
+    version: '1.5.35',
+    rules: 'modern'
+};