+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# 1.5.35 Release:
+## Update Notes:
+- This is not a V14 update. An update for V14 is still a work in progress.
+## New Modern Content:
+- Unarmed Strike (Monk, modern)
+- Animate Objects
+## New Legacy Content:
+- Martial Arts
+- Slow Fall
+- Animate Objects (Thanks BinaryScrub)
+## Bug Fixes:
+- Iron Chin
+- Fisticuffs
+- Unarmed Strike (Pugilist)
+- Savage Attacker (modern and legacy)
+- Ki
+- Stunning Strike (legacy)
+- Mantle of Inspiration
+- Eldritch Invocations: Agonizing Blast
+
 # 1.5.34 Release:
 ## Update Notes:
 - This is not a V14 update. An update for V14 is still a work in progress.